move appModel to root dir
