+++ packages/editor/__tests__/model/componentModel.spec.ts
@@ -1,10 +1,10 @@
-import { ApplicationModel } from '../../src/operations/AppModel/AppModel';
+import { ApplicationModel } from '../../src/AppModel/AppModel';
 import {
   ComponentId,
   ComponentType,
   SlotName,
   TraitType,
-} from '../../src/operations/AppModel/IAppModel';
+} from '../../src/AppModel/IAppModel';
 import { AppSchema } from './schema';
 
 describe('ComponentModel test', () => {
