+++ packages/editor/src/operations/leaf/component/adjustComponentOrderLeafOperation.ts
@@ -1,6 +1,6 @@
 import { ApplicationComponent } from '@sunmao-ui/core';
-import { ApplicationModel } from '../../AppModel/AppModel';
-import { ComponentId } from '../../AppModel/IAppModel';
+import { ApplicationModel } from '../../../AppModel/AppModel';
+import { ComponentId } from '../../../AppModel/IAppModel';
 import { BaseLeafOperation } from '../../type';
 export type AdjustComponentOrderLeafOperationContext = {
   orientation: 'up' | 'down';
