+++ packages/editor/src/operations/leaf/component/pasteComponentLeafOperation.ts
@@ -1,6 +1,6 @@
 import { ApplicationComponent } from '@sunmao-ui/core';
-import { ApplicationModel } from '../../AppModel/AppModel';
-import { ComponentId, IComponentModel, SlotName } from '../../AppModel/IAppModel';
+import { ApplicationModel } from '../../../AppModel/AppModel';
+import { ComponentId, IComponentModel, SlotName } from '../../../AppModel/IAppModel';
 import { BaseLeafOperation } from '../../type';
 
 export type PasteComponentLeafOperationContext = {
